remove DynEncode method and replace it with a normal function
