+++ pkg/sr/clientopt.go
