+++ pkg/sr/serde.go
@@ -20,77 +20,6 @@ var (
 	ErrBadHeader = errors.New("5 byte header for value is missing or does not have 0 magic byte")
 )
 
-type (
-	// EncodingOpt is an option to configure the behavior of Serde.Encode and
-	// Serde.Decode.
-	EncodingOpt interface {
-		serdeOrEncodingOpt()
-		apply(*tserde)
-	}
-	encodingOpt struct{ fn func(*tserde) }
-
-	// SerdeOpt is an option to configure a Serde.
-	SerdeOpt interface {
-		serdeOrEncodingOpt()
-		apply(*Serde)
-	}
-	serdeOpt struct{ fn func(serde *Serde) }
-
-	// SerdeOrEncodingOpt is either a SerdeOpt or EncodingOpt.
-	SerdeOrEncodingOpt interface {
-		serdeOrEncodingOpt()
-	}
-)
-
-func (o serdeOpt) serdeOrEncodingOpt() { /* satisfy interface */ }
-func (o serdeOpt) apply(s *Serde)      { o.fn(s) }
-
-func (o encodingOpt) serdeOrEncodingOpt() { /* satisfy interface */ }
-func (o encodingOpt) apply(t *tserde)     { o.fn(t) }
-
-// EncodeFn allows Serde to encode a value.
-func EncodeFn(fn func(any) ([]byte, error)) EncodingOpt {
-	return encodingOpt{func(t *tserde) { t.encode = fn }}
-}
-
-// AppendEncodeFn allows Serde to encode a value to an existing slice. This
-// can be more efficient than EncodeFn; this function is used if it exists.
-func AppendEncodeFn(fn func([]byte, any) ([]byte, error)) EncodingOpt {
-	return encodingOpt{func(t *tserde) { t.appendEncode = fn }}
-}
-
-// DecodeFn allows Serde to decode into a value.
-func DecodeFn(fn func([]byte, any) error) EncodingOpt {
-	return encodingOpt{func(t *tserde) { t.decode = fn }}
-}
-
-// GenerateFn returns a new(Value) that can be decoded into. This function can
-// be used to control the instantiation of a new type for DecodeNew.
-func GenerateFn(fn func() any) EncodingOpt {
-	return encodingOpt{func(t *tserde) { t.gen = fn }}
-}
-
-// Index attaches a message index to a value. A single schema ID can be
-// registered multiple times with different indices.
-//
-// This option supports schemas that encode many different values from the same
-// schema (namely, protobuf). The index into the schema to encode a
-// particular message is specified with `index`.
-//
-// NOTE: this option must be used for protobuf schemas.
-//
-// For more information, see where `message-indexes` are described in:
-//
-//	https://docs.confluent.io/platform/current/schema-registry/serdes-develop/index.html#wire-format
-func Index(index ...int) EncodingOpt {
-	return encodingOpt{func(t *tserde) { t.index = index }}
-}
-
-// Header defines the SerdeHeader used to encode and decode the message header.
-func Header(header SerdeHeader) SerdeOpt {
-	return serdeOpt{func(s *Serde) { s.h = header }}
-}
-
 type tserde struct {
 	id           uint32
 	exists       bool
@@ -286,63 +215,37 @@ func tserdeMapClone(m map[int]tserde, at int, index []int) map[int]tserde {
 	return dup
 }
 
-// Encode encodes a value with a prepended header according to the configured
-// SerdeHeader. If EncodeFn was not registered, this returns ErrNotRegistered.
+// Encode encodes a value and prepends the header according to the configured
+// SerdeHeader. If EncodeFn was not used, this returns ErrNotRegistered.
 func (s *Serde) Encode(v any) ([]byte, error) {
 	return s.AppendEncode(nil, v)
 }
 
-// AppendEncode encodes a value with a prepended header according to the
+// AppendEncode encodes a value and prepends the header according to the
 // configured SerdeHeader, appends it to b and returns b. If EncodeFn was not
 // registered, this returns ErrNotRegistered.
 func (s *Serde) AppendEncode(b []byte, v any) ([]byte, error) {
 	// Load tserde based on the registered type.
 	t := s.loadTypes()[reflect.TypeOf(v)]
-	return s.encode(b, v, t)
-}
 
-// DynEncode encodes a value with a prepended header according to the configured
-// SerdeHeader. If EncodeFn was not registered using the supplied id and index,
-// this returns ErrNotRegistered.
-func (s *Serde) DynEncode(v any, id int, index []int) ([]byte, error) {
-	return s.DynAppendEncode(nil, v, id, index)
-}
-
-// DynAppendEncode encodes a value with a prepended header according to the
-// configured SerdeHeader, appends it to b and returns b. If EncodeFn was not
-// registered using the supplied id and index, this returns ErrNotRegistered.
-func (s *Serde) DynAppendEncode(b []byte, v any, id int, index []int) ([]byte, error) {
-	// Load tserde based on the supplied ID.
-	t := s.loadIDs()[id]
-	// Traverse to the right index, if index is supplied.
-	for _, i := range index {
-		if len(t.subindex) <= i {
-			return nil, ErrNotRegistered
-		}
-		t = t.subindex[i]
-	}
-	return s.encode(b, v, t)
-}
-
-func (s *Serde) encode(b []byte, v any, t tserde) ([]byte, error) {
 	// Check if we loaded a valid tserde.
 	if !t.exists || (t.encode == nil && t.appendEncode == nil) {
 		return nil, ErrNotRegistered
 	}
 
-	b, err := s.header().AppendEncode(b, int(t.id), t.index)
-	if err != nil {
-		return nil, err
+	appendEncode := t.appendEncode
+	if appendEncode == nil {
+		// Fallback to t.encode.
+		appendEncode = func(b []byte, v any) ([]byte, error) {
+			encoded, err := t.encode(v)
+			if err != nil {
+				return nil, err
+			}
+			return append(b, encoded...), nil
+		}
 	}
 
-	if t.appendEncode != nil {
-		return t.appendEncode(b, v)
-	}
-	encoded, err := t.encode(v)
-	if err != nil {
-		return nil, err
-	}
-	return append(b, encoded...), nil
+	return AppendEncode(b, v, int(t.id), t.index, s.header(), appendEncode)
 }
 
 // MustEncode returns the value of Encode, panicking on error. This is a
@@ -359,26 +262,6 @@ func (s *Serde) MustEncode(v any) []byte {
 // This is a shortcut for if your encode function cannot error.
 func (s *Serde) MustAppendEncode(b []byte, v any) []byte {
 	b, err := s.AppendEncode(b, v)
-	if err != nil {
-		panic(err)
-	}
-	return b
-}
-
-// MustDynEncode returns the value of DynEncode, panicking on error. This is a
-// shortcut for if your encode function cannot error.
-func (s *Serde) MustDynEncode(v any, id int, index []int) []byte {
-	b, err := s.DynEncode(v, id, index)
-	if err != nil {
-		panic(err)
-	}
-	return b
-}
-
-// MustDynAppendEncode returns the value of DynAppendEncode, panicking on error.
-// This is a shortcut for if your encode function cannot error.
-func (s *Serde) MustDynAppendEncode(b []byte, v any, id int, index []int) []byte {
-	b, err := s.DynAppendEncode(b, v, id, index)
 	if err != nil {
 		panic(err)
 	}
@@ -445,6 +328,28 @@ func (s *Serde) decodeFind(b []byte) ([]byte, tserde, error) {
 	}
 
 	return b, t, nil
+}
+
+// Encode encodes a value and prepends the header. If the encoding function
+// fails, this returns an error.
+func Encode(v any, id int, index []int, h SerdeHeader, enc func(any) ([]byte, error)) ([]byte, error) {
+	return AppendEncode(nil, v, id, index, h, func(b []byte, val any) ([]byte, error) {
+		encoded, err := enc(val)
+		if err != nil {
+			return nil, err
+		}
+		return append(b, encoded...), nil
+	})
+}
+
+// AppendEncode encodes a value and prepends the header, appends it to b and
+// returns b. If the encoding function fails, this returns an error.
+func AppendEncode(b []byte, v any, id int, index []int, h SerdeHeader, enc func([]byte, any) ([]byte, error)) ([]byte, error) {
+	b, err := h.AppendEncode(b, id, index)
+	if err != nil {
+		return nil, err
+	}
+	return enc(b, v)
 }
 
 // SerdeHeader encodes and decodes a message header.
